backend/requirements.txt
# backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
motor==3.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
google-cloud-vision==3.4.4
taxjar
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
pytest==7.4.4
httpx==0.26.0
bcrypt==4.1.2
python-magic==0.4.27
aiofiles==23.2.1
pymongo==4.6.1